requirements.txt
absl-py==1.4.0
astor==0.8.1
bert-serving-client==1.10.0
bert-serving-server==1.10.0
Bottleneck @ file:///opt/conda/conda-bld/bottleneck_1657175564434/work
certifi @ file:///croot/certifi_1671487769961/work/certifi
cffi @ file:///tmp/build/80754af9/cffi_1636541929935/work
charset-normalizer==3.3.2
cryptography @ file:///croot/cryptography_1677533068310/work
docopt==0.6.2
et-xmlfile==1.1.0
gast==0.5.4
google-pasta==0.2.0
GPUtil==1.4.0
grpcio==1.57.0
h5py==3.8.0
idna==3.6
importlib-metadata==6.7.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
Markdown==3.4.4
MarkupSafe==2.1.3
mkl-fft==1.3.1
mkl-random @ file:///tmp/build/80754af9/mkl_random_1626179032232/work
mkl-service==2.4.0
numexpr @ file:///tmp/build/80754af9/numexpr_1618856527787/work
numpy==1.21.6
openpyxl==3.1.2
pandas==1.3.5
pipreqs==0.4.13
protobuf==3.19.0
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
PyMySQL @ file:///tmp/build/80754af9/pymysql_1610482900657/work
python-dateutil @ file:///tmp/build/80754af9/python-dateutil_1626374649649/work
pytz==2023.3
pyzmq==25.1.1
requests==2.31.0
six @ file:///tmp/build/80754af9/six_1644875935023/work
tensorboard==1.14.0
tensorflow==1.14.0
tensorflow-estimator==1.14.0
termcolor==2.3.0
typing_extensions==4.7.1
urllib3==2.0.7
Werkzeug==2.2.3
wrapt==1.15.0
yarg==0.1.9
zipp==3.15.0
